Remove Reps bundles
