+++ src/devtools/client/debugger/packages/devtools-reps/src/reps/document.js
@@ -50,12 +50,7 @@ function getTitle(grip) {
 
 // Registration
 function supportsObject(object, noGrip = false) {
-  if (noGrip === true || !isGrip(object)) {
-    return false;
-  }
-
-  const type = getGripType(object, noGrip);
-  return object.preview && type === "HTMLDocument";
+  return object.className() === "HTMLDocument";
 }
 
 // Exports from this module
