+++ src/devtools/client/debugger/packages/devtools-reps/src/reps/stylesheet.js
@@ -46,12 +46,8 @@ function getLocation(grip) {
 }
 
 // Registration
-function supportsObject(object, noGrip = false) {
-  if (noGrip === true || !isGrip(object)) {
-    return false;
-  }
-
-  return getGripType(object, noGrip) == "CSSStyleSheet";
+function supportsObject(object) {
+  return object.className() == "CSSStyleSheet"
 }
 
 // Exports from this module
