+++ src/devtools/client/debugger/packages/devtools-reps/src/reps/window.js
@@ -58,11 +58,7 @@ function getLocation(object) {
 
 // Registration
 function supportsObject(object, noGrip = false) {
-  if (noGrip === true || !isGrip(object)) {
-    return false;
-  }
-
-  return object.preview && getGripType(object, noGrip) == "Window";
+  return object.className() == "Window";
 }
 
 // Exports from this module
