+++ src/devtools/client/webconsole/components/Output/GripMessageBody.js
@@ -15,7 +15,7 @@ const {
 } = require("devtools/client/webconsole/utils/object-inspector");
 const actions = require("devtools/client/webconsole/actions/index");
 
-const reps = require("devtools/client/shared/components/reps/reps");
+const reps = require("devtools/client/debugger/packages/devtools-reps/src");
 const { objectInspector, MODE } = reps;
 
 GripMessageBody.displayName = "GripMessageBody";
@@ -89,8 +89,8 @@ function GripMessageBody(props) {
 // Regular expression that matches the allowed CSS property names.
 const allowedStylesRegex = new RegExp(
   "^(?:-moz-)?(?:background|border|box|clear|color|cursor|display|float|font|line|" +
-    "margin|padding|text|transition|outline|white-space|word|writing|" +
-    "(?:min-|max-)?width|(?:min-|max-)?height)"
+  "margin|padding|text|transition|outline|white-space|word|writing|" +
+  "(?:min-|max-)?width|(?:min-|max-)?height)"
 );
 
 // Regular expression that matches the forbidden CSS property values.
